azure-riot-function (OLD)/requirements.txt
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
mysql-connector==2.2.9
portalocker==2.10.1
prettytable==3.12.0
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
